refactor: logic
Now all the logic of cart component is going on using a fetcher,
so that it doesnt navigate whenever a change happens.
Also, had to change the logic of shoppingCartApi, as it was only
comparing items by its item id, but because some items can have a
size property, they're actually different objects, and API couldn't
undrestand it. So now, that I refactored it, it can differentiate
items by its size even if they have the same item id